+++ test/recurly/coupon_test.php
@@ -31,4 +31,18 @@ class Recurly_CouponTest extends UnitTestCase
     $this->assertEqual($xml,
       "<?xml version=\"1.0\"?>\n<coupon><coupon_code>fifteen-off</coupon_code><name>$15 Off</name><discount_type>dollar</discount_type><discount_in_cents><USD>1500</USD></discount_in_cents></coupon>\n");
   }
+
+  public function testXmlWithPlans()
+  {
+    $coupon = new Recurly_Coupon();
+    $coupon->coupon_code = 'fifteen-off';
+    $coupon->name = '$15 Off';
+    $coupon->discount_type = 'dollar';
+    $coupon->discount_in_cents->addCurrency('USD', 1500);
+    $coupon->plan_codes = array('gold', 'monthly');
+
+    $xml = $coupon->xml();
+    $this->assertEqual($xml,
+      "<?xml version=\"1.0\"?>\n<coupon><coupon_code>fifteen-off</coupon_code><name>$15 Off</name><discount_type>dollar</discount_type><discount_in_cents><USD>1500</USD></discount_in_cents><plan_codes><plan_code>gold</plan_code><plan_code>monthly</plan_code></plan_codes></coupon>\n");
+  }
 }